pyproject.toml
[tool.poetry]
name = "fastapiproject"
version = "0.1.0"
description = ""
authors = ["kang9805 <zzionkk@naver.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.116.1"
uvicorn = "^0.35.0"
orjson = "3.10.12"
tortoise-orm = {version = "0.23.0", extras = ["asyncmy"]}
cryptography = "44.0.0"
aerich = "0.8.1"
tomlkit = "0.13.2"
pydantic-settings = "2.7.1"


[tool.poetry.group.dev.dependencies]
black = "24.10.0"
ruff = "0.8.2"
mypy = "1.13.0"
pytest = "8.3.4"
coverage = "7.6.9"
pytest-asyncio = "0.25.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.ruff]
target-version = "py313"


[tool.mypy]
plugins = ["pydantic.mypy"]
python_version = 3.13
strict = true
exclude = [
    "1일차/.*",
    "2일차/.*"
]

[tool.coverage.run]
omit = ["*/test_*.py"]


[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"

